doc-ai/form-parser/cloudrun/requirements.txt
google-auth==2.6.3
google-cloud-documentai==1.2.1
google-cloud-storage==2.0.0
pandas
google-cloud-bigquery
google-cloud-pubsub